Fix issue with transitioning to finish at Asherons Castle
Asheron's Castle is right on the edge of a landblock and if you run
forward it does not know you are at the destination and stays in the
state `UseEmpyreanPortalspaceActivator`. This is mostly fixed by using
`IntoPortal` instead of `ExitPortal` but I also switched to using the
larger landblock instead of landcell.

- Use landblock instead of landcell
- Use IntoPortal instead of ExitPortal