Add `#jump on me` command
When jumping I want all characters to bunch up on top of me. I frequently set follow to 0.1, wait for them to get on top of me and then return follow to 1.5. Instead there should be an On me button in the jump panel that causes the characters to follow at 0.1 and when they are within 0.1-0.2 away return their follow to the previous value.

Fix #116